+++ CHANGELOG.md
@@ -110,6 +110,7 @@
   * `stubbles\lang\reflect()` can now also reflect functions
   * `stubbles\lang\reflect()` now throws a `stubbles\lang\exception\IllegalArgumentException` if the value can not be reflected
   * added `stubbles\lang\ensureCallable()`
+  * added `stubbles\peer\IpAddress`
 
 
 3.5.3 (2014-05-07)